requirements.txt
mysql-connector
dialogflow
requests
flask
flask_cors
